Refactor to use deploy command as plugin
