+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed 
+- [vtex deploy] Refactor to use as oclif plugin.
+
 ## [2.112.0-beta] - 2020-09-15
 ### Added
 - [installer] New deploy method using `oclif-dev pack` and `oclif-dev publish`.
