+++ src/modules/apps/deprecate.ts
@@ -6,7 +6,7 @@ import { SessionManager } from '../../api/session/SessionManager'
 import { TelemetryCollector } from '../../lib/telemetry/TelemetryCollector'
 import { parseLocator } from '../../api/locator'
 import log from '../../api/logger'
-import { returnToPreviousAccount, switchAccount } from '../auth/switch'
+import { returnToPreviousAccount, switchAccount } from '../../api/modules/auth/switch'
 import { promptConfirm } from '../../api/modules/prompts'
 
 let originalAccount
