+++ src/modules/apps/publish.ts
@@ -13,7 +13,7 @@ import { promptConfirm } from '../../api/modules/prompts'
 import { checkBuilderHubMessage, showBuilderHubMessage } from '../../api/modules/utils'
 import { SessionManager } from '../../api/session/SessionManager'
 import * as conf from '../../api/conf'
-import { returnToPreviousAccount, switchAccount } from '../auth/switch'
+import { returnToPreviousAccount, switchAccount } from '../../api/modules/auth/switch'
 import { runYarnIfPathExists } from '../utils'
 
 const buildersToRunLocalYarn = ['node', 'react']
